Move files to main folder
